Add files via upload
We define a palindrome as a sequence of letters or digits that is the same backward as forward.
A few examples of palindromes are:
“Never odd or even”
“Was it a cat I saw?”
“A man, a plan, a canal --- Panama!”
Note that palindromes are case-insensitive (do not care if a character is uppercase or lowercase), and punctuations and white spaces are ignored (all the white spaces and punctuations are treated as if they were not there). Only letters and digits are considered in a palindrome. You may find it helpful to look at the documentation of the Character class of the Java API (google “java character”) and use one of its methods. Write an application (by creating a classed called PalindromeTester) that prompts user to enter a palindrome and reads the sentence typed in, and determines if it is a palindrome and displays the conclusion on screen; then it asks user if they want to test another palindrome (y/n?), if user types in ‘y’ then repeat the previous step, and this is repeated until user types in ‘n’.