Add files via upload
Create a Sphere class to represent a sphere with instance data of the diameter. Write appropriate constructor, getter and setter methods. Write methods to return the volume and surface area of the sphere. The formulas are:
Volume =4/3πr^3
SurfaceArea = 4πr^2
Write a toString method that returns a string description of the sphere. Also create a driver class called MultiSphere whose main method instantiates and updates three Sphere objects.